Changed x and y from int to double

+++ src/compsciproject2023/Enemies.java
@@ -12,7 +12,7 @@ package compsciproject2023;
 public class Enemies extends Sprite {
     int StrengthE, Health;
     
-    public Enemies(int width, int height, String imgpath, int stength, int health, int x, int y, int Range) {
+    public Enemies(int width, int height, String imgpath, int stength, int health, double x, double y, int Range) {
         super(width, height, imgpath);
         this.StrengthE = stength;
         this.Health = health;
@@ -21,9 +21,9 @@ public class Enemies extends Sprite {
         setYSprite(y);
     }
     
-    private void Enemyattack(int playerx, int playery){
-        int playerdistx = getXSprite()-playerx;
-        int playerdisty = getYSprite()-playery;
+    private void Enemyattack(double playerx, double playery){
+        double playerdistx = getXSprite()-playerx;
+        double playerdisty = getYSprite()-playery;
         
         if ((playerdistx)*(playerdistx)<(getRange())*(getRange())) {
